requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
python-telegram-bot==21.10
vercel>=0.5.0
